% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_run.R
\name{analyze_run}
\alias{analyze_run}
\title{Segmentation of the running route according to activity}
\usage{
analyze_run(gps, smoothing = FALSE)
}
\arguments{
\item{gps}{The GPS data of the running route (longitude, latitude, time, elevation, ...). The longitude and latitude variables must be named "lon" and "lat", respectfully.}

\item{smoothing}{Indicator of whether the results should be calculated on smoothed data or not. It is set to FALSE by default.}
}
\value{
A list containing:
 \itemize{
   \item moving_ind indexes of GPS data points with velocity greater than 1 m/s,
   \item resting_ind indexes of GPS data points with velocity below 1 m/s.
   \item time_div data frame
     \itemize{
       \item active_time total duration of the running route,
       \item moving_time duration of the runner traveling faster than 1 m/s (3.6 km/h),
       \item t_mov_rel moving time relative to total time,
       \item resting_time duration of the runner traveling slower than 1 m/s (3.6 km/h),
       \item t_rest_rel resting time relative to total time,
       \item time_asc total time spent ascending,
       \item t_asc_rel time spent ascending relative to total time,
       \item time_desc total time spent descending,
       \item t_desc_rel time spent descending relative to total time,
       \item time_flat total time spent on flat,
       \item t_flat_rel time spent on flat relative to total time,
       \item time_NA total time with undefined (NA) grade because of smoothing,
       \item t_NA_rel time with undefined (NA) grade because of smoothing relative to total time.
     }
   \item pace data frame
     \itemize{
       \item avg_pace average pace for the entire activity,
       \item avg_moving_pace average pace for the moving time,
       \item best_pace fastest achieved pace,
       \item worst_pace slowest achieved pace,
     }
 }
}
\description{
This function determines the total time, moving time and resting time on a running route described with GPS data.
Additionally, the function calculates the time spent ascending, descending and on the flat.
}
\examples{
load(system.file("extdata", "gpsHR.Rda", package="runneR"))
segments <- analyze_run(gpsHR)
}
\keyword{GPS}
\keyword{analysis,}
\keyword{data}
\keyword{moving}
\keyword{running,}
\keyword{time}

