% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{bmi}
\alias{bmi}
\title{Calculates Body Mass Index}
\usage{
bmi(height, weight, units = "english")
}
\arguments{
\item{height}{Height of individual of numeric vector. In inches if english units and meters if metric units.}

\item{weight}{Weight of individual of numeric vector. In pounds if english units and kilograms if metric units.}

\item{units}{Character vector representing the units convention used for height and weight. Either "english" or "metric". Default is "english".}
}
\value{
numeric vector representing the BMI of an individual based on their height and weight
}
\description{
Calculates Body Mass Index
}
\examples{
bmi(72,1.7,"metric")
bmi(66,140)
}

