% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body_fat.R
\name{body_fat}
\alias{body_fat}
\title{Calculates Body Fat Percentage}
\usage{
body_fat(weight, gender, waist, wrist = 0, hip = 0, forearm = 0,
  units = "english")
}
\arguments{
\item{weight}{Weight of individual of numeric vector. In pounds if english units and kilograms if metric units.}

\item{waist}{Waist measurement at navel of individual of numeric vector. In inches if english units and centimeters if metric units.}

\item{wrist}{Wrist measurement of individual of numeric vector. In inches if english units and centimeters if metric units.}

\item{hip}{Hip measurement of individual at fullest point of numeric vector. In inches if english units and centimeters if metric units.}

\item{forearm}{Forearm measurement of individual at fullest point of numeric vector. In inches if english units and centimeters if metric units.}

\item{units}{Character vector representing the units convention used for variables. Either "english" or "metric". Default is "english".}

\item{gendsnter}{"male" for male and "female" for female.}
}
\value{
numeric vector representing the BMI of an individual based on their height and weight
}
\description{
Calculates Body Fat Percentage
}
\examples{
body_fat(152, "male",30, 6.5, 34, 10, "english")
}

