% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calories_burned.R
\name{calories_burned}
\alias{calories_burned}
\title{Calculates Calories Burned From Exercise}
\usage{
calories_burned(age, heart_rate, weight, time, gender, units = "english")
}
\arguments{
\item{age}{Age of individual of numeric vector.}

\item{heart_rate}{Heart rate (bpm) of individual of numeric vector.}

\item{weight}{Weight of individual of numeric vector. In pounds if english units and kilograms if metric units.}

\item{time}{Time spent running of individual of numeric vector. In minutes.}

\item{units}{Character vector representing the units convention used for weight. Either "english" or "metric". Default is "english".}

\item{gendser}{"male" for male and "female" for female.}
}
\value{
numeric vector representing the BMI of an individual based on their height and weight
}
\description{
Calculates Calories Burned From Exercise
}
\examples{
calories_burned(19, 130, 152, 60, "male", "metric")
}

