% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_activity.R
\name{cb_activity}
\alias{cb_activity}
\title{Estimate calorie burn based on running speed}
\usage{
cb_activity(gender, age, m, h, gps)
}
\arguments{
\item{gender}{String, either "male" or "female"}

\item{age}{Age of the person in years}

\item{m}{Weight in kg}

\item{h}{Height in cm}

\item{gps}{the GPS data of the running route (longitude, latitude, time, elevation, ...). The longitude and latitude variables must be named "lon" and "lat", respectfully.}
}
\value{
list containing
 \itemize{
   \item netCB Estimated net calories burned during the run,
   \item gps the input data frame with additional columns of net calorie burn values (netCB), distances (d) and time differences (delta_time).
 }
}
\description{
This function estimates the amount of calories that you burn while running activity based on the MET (Metabolic Equivalent) data for physical activities from "The Compendium of Physical Activities Tracking Guide" by B. E. Ainsworth.
All input parameters are mandatory.
}
\examples{
load(system.file("extdata", "gpsHR.Rda", package="runneR"))
activity_cal_burn <- cb_activity(gender = "male", age = 41, m = 80, h = 180, gpsHR)
}
\keyword{GPS}
\keyword{analysis,}
\keyword{burn}
\keyword{calorie}
\keyword{data}
\keyword{running,}

