% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_hr.R
\name{cb_hr}
\alias{cb_hr}
\title{Estimate calorie burn based on heart rate}
\usage{
cb_hr(gender, age, m, h, RHR, HR, gps)
}
\arguments{
\item{gender}{String, either "male" or "female"}

\item{age}{Age of the person in years}

\item{m}{Weight in kg}

\item{h}{Height in cm}

\item{RHR}{Resting heart rate}

\item{HR}{heart rate time series data for the run}

\item{gps}{the GPS data of the running route (longitude, latitude, time, elevation, ...). The longitude and latitude variables must be named "lon" and "lat", respectfully.}
}
\value{
list containing
 \itemize{
   \item netCB Estimated net calories burned during the run,
   \item gps the input data frame with additional columns of net calorie burn values (netCB), distances (d) and time differences (delta_time).
 }
}
\description{
This heart rate based calorie burn calculator provides an estimate of the
rate at which you are burning calories during aerobic (i.e. cardiorespiratory)
exercise, based on your average heart rate while performing the exercise.
All input parameters are mandatory.
}
\examples{
load(system.file("extdata", "gpsHR.Rda", package="runneR"))
hr_cal_burn <- cb_hr(gender = "male", age = 41, m = 80, h = 180, RHR = 60, gpsHR$hr, gpsHR)
}
\keyword{GPS}
\keyword{analysis,}
\keyword{burn}
\keyword{calorie}
\keyword{data}
\keyword{running,}

