% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_running.R
\name{cb_running}
\alias{cb_running}
\title{Estimate calorie burn while running any given distance}
\usage{
cb_running(age, m, h, RHR, treadmill = FALSE, gps)
}
\arguments{
\item{age}{Age of the person in years}

\item{m}{Weight in kg}

\item{h}{Height in cm}

\item{RHR}{Resting heart rate in beats/min}

\item{treadmill}{Indicator of whether the running was performed on a treadmill or not (TRUE or FALSE)}

\item{gps}{the GPS data of the running route containing time, longitude, latitude and elevation. The longitude, latitude and elevation variables must be named "lon", "lat" and "ele", respectfully.}
}
\value{
list containing
 \itemize{
   \item netCB Estimated net calories burned during the run,
   \item gps the input data frame with additional columns of net calorie burn values (netCB), distances (d) and time differences (delta_time).
 }
}
\description{
This running calorie burn calculator estimates the calories that you burn while running any given distance.
The calculator takes into consideration the grade of the running surface that you are on (i.e. the incline or decline),
whether you are running on a treadmill or not, and your fitness level.
}
\examples{
load(system.file("extdata", "gpsHR.Rda", package="runneR"))
running_cal_burn <- cb_running(age = 41, m = 80, h = 180, RHR = 60, treadmill = FALSE, gpsHR)
}
\keyword{GPS}
\keyword{analysis,}
\keyword{burn}
\keyword{calorie}
\keyword{data}
\keyword{running,}

