% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{check_k}
\alias{check_k}
\title{Checks if k is valid}
\usage{
check_k(k, n, var)
}
\arguments{
\item{k}{\code{integer} vector}

\item{n}{\code{integer} length of code{x}}

\item{var}{\code{character} name of the variable}
}
\description{
Checks if k is valid
}
