% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_log.R
\name{create_log}
\alias{create_log}
\title{Create a running log to track running data}
\usage{
create_log(units = "english", date, name, age, gender, height, weight, time,
  distance, heartrate = NA)
}
\arguments{
\item{units}{"english" or "metric"}

\item{date}{User's date. Should be in "mm/dd/yyyy" format, of the class character.}

\item{name}{User's name. Should be in}

\item{age}{User's age. Should be in numeric format.}

\item{gender}{User's gender. Should be either "male" or "female"}

\item{height}{User's height. Should be in numeric format. If units = "english", should be in inches. If units = "metric", should be in centimeters.}

\item{weight}{User's weight. Should be in numeric format.}

\item{time}{The number of minutes the user ran the distance in. Should be in numeric format.}

\item{distance}{The distance the user ran. Should be in numeric format.}

\item{heartrate}{Optional parameter - user's heartrate during workout. Should be in numeric format.}

\item{filename}{The name under which you want the file saved. Should include .rds extension.}
}
\description{
Creates a new log for the user's running data. Using user info and measurements,
             creates a data frame with the info entered along with calories burned, bmi, and body_fat.
             Saves the data as a .RData file under the user's chosen file name.
}
\examples{
create_log("english", "09/30/2016", "Anna", 18, "female", 57, 128, 10, 1)
create_log("metric", "09/30/2016", "Jon", 19, "male", 57, 128, 10, 1, 72)
}

