% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_unresolved_at}
\alias{.check_unresolved_at}
\title{Resolves at argument}
\usage{
.check_unresolved_at(x, at)
}
\arguments{
\item{x}{(\code{data.frame})}

\item{at}{(\code{integer}, \code{Date}, \code{POSIXt}, \code{character})\cr
Indices at which to evaluate windows. Output length equals \code{length(at)}
instead of \code{length(x)}. A single time-interval string (e.g. \code{"month"})
generates a regular sequence over the range of \code{idx}.}
}
\value{
resolved \code{at}
}
\description{
Resolves \code{at} argument passed to the \code{runner} -
checks if the argument has a valid value. If argument is a single character
matching column name in the \code{x} then is replaced with the value \code{x[[at]]}
}
\keyword{internal}
