% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_unresolved_difftime}
\alias{.check_unresolved_difftime}
\title{Resolves time difference argument}
\usage{
.check_unresolved_difftime(x, k)
}
\arguments{
\item{x}{(\code{data.frame})}

\item{k}{(\code{integer} or \code{character})\cr
Window size. Single value or vector of \code{length(x)}. Omit for cumulative
windows. Accepts time-interval strings (e.g. \code{"5 days"}) when \code{idx} is set.}
}
\value{
resolved \code{idx}
}
\description{
Resolves \code{at} argument passed to the \code{runner} (\code{k} or \code{lag})
checks if the argument has a valid value. If argument is a single character
matching column name in the \code{x} then is replaced with the value \code{x[[idx]]}
}
\keyword{internal}
