% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_datetime_valid}
\alias{.is_datetime_valid}
\title{Validate date time character}
\usage{
.is_datetime_valid(x)
}
\arguments{
\item{(`character`)}{can be anything but suppose to be a character.}
}
\value{
\code{logical(1)} denoting if all elements of the character vectors are valid
}
\description{
Checks if the character is a valid date time string
}
\keyword{internal}
