% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.k_by}
\alias{.k_by}
\title{Converts k and lag from time-unit-interval to int}
\usage{
.k_by(k, idx, param)
}
\arguments{
\item{k}{(\code{integer} or \code{character})\cr
Window size. Single value or vector of \code{length(x)}. Omit for cumulative
windows. Accepts time-interval strings (e.g. \code{"5 days"}) when \code{idx} is set.}

\item{idx}{(\code{integer}, \code{Date}, \code{POSIXt})\cr
Sorted index of observations. When set, \code{k} and \code{lag} refer to index
distance rather than element count. Must be same length as \code{x}.}

\item{param}{name of the parameter to be printed in error message}
}
\description{
Converts k and lag from time-unit-interval to int
}
\examples{
k <- "1 month"
idx <- seq(
  as.POSIXct("2019-01-01 03:02:01"),
  as.POSIXct("2020-01-01 03:02:01"),
  by = "month"
)
k_difftime <- runner:::.k_by(k, idx, param = "k")
idx - k_difftime
}
\keyword{internal}
