% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.reformat_k}
\alias{.reformat_k}
\title{Formats time-unit-interval to valid for runner}
\usage{
.reformat_k(k, only_positive = TRUE)
}
\arguments{
\item{k}{(k or lag) object from runner to be formatted}

\item{only_positive}{for \code{k} is \code{TRUE},
for \code{lag} is \code{FALSE}}
}
\description{
Formats time-unit-interval to valid for runner. User specifies \code{k} as
positive number but this means that this interval needs to be substracted
from \code{idx} - because windows length extends window backwards in time.
The same situation for lag.
}
\examples{
runner:::.reformat_k("1 days")
runner:::.reformat_k("day")
runner:::.reformat_k("10 days")
runner:::.reformat_k("-10 days", only_positive = FALSE)
runner:::.reformat_k(c("-10 days", "2 months"), only_positive = FALSE)
}
\keyword{internal}
