% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.resolve_arg}
\alias{.resolve_arg}
\title{Resolves at argument}
\usage{
.resolve_arg(x, arg)
}
\arguments{
\item{x}{(\code{data.frame})}
}
\value{
resolved \code{at}
}
\description{
Resolves argument passed to the \code{runner} -
checks if the argument has a valid value. If argument is a single character
matching column name in the \code{x} then is replaced with the value \code{x[[arg]]}
}
\keyword{internal}
