% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.seq_at}
\alias{.seq_at}
\title{Creates sequence for at as time-unit-interval}
\usage{
.seq_at(at, idx)
}
\arguments{
\item{at}{object from runner}

\item{idx}{object from runner}
}
\description{
Creates sequence for at as time-unit-interval
}
\keyword{internal}
