% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.this_group}
\alias{.this_group}
\title{Access group data in mutate}
\usage{
.this_group(x)
}
\arguments{
\item{x}{(\code{data.frame})\cr
object which can be \code{grouped_df} in special case.}
}
\value{
data.frame filtered by current \code{dplyr::groups()}
}
\description{
Access group data in \code{dplyr::mutate} after
\code{dplyr::group_by}.
Function created because data available in \code{dplyr::group_by \%>\% mutate}
scheme is not filtered by group - in mutate function \code{.} is still initial
dataset. This function creates \code{data.frame} using \code{dplyr::groups}
information.
}
\keyword{internal}
