% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dummy_min}
\alias{dummy_min}
\alias{dummy_hour}
\alias{dummy_day}
\title{Dummy data.}
\format{
A data frame with 100000 rows and 2 columns
\describe{
  \item{min}{minute - irregulary spaced}
  \item{val1}{random walk data}
  \item{val2}{random walk data - 2 * val1 + epsilon}
  \item{val3}{random walk data - -1.5 * val1 + 2epsilon}
  \item{val4}{random walk data - 0.7 * val2 + epsilon}
}

A data frame with 10000 rows and 2 columns (subset of `dummy_min`)

A data frame with 1000 rows and 2 columns (subset of `dummy_min`)
}
\source{
internal

internal

internal
}
\usage{
dummy_min

dummy_hour

dummy_day
}
\description{
A dataset containing dummy time-serie in minute interval.

A dataset containing dummy time-serie in hourly interval.

A dataset containing dummy time-serie in daily interval.
}
\examples{
dummy_min
}
\keyword{datasets}
