% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fill_along}
\alias{fill_along}
\title{Repeat non-NA (fill further)}
\usage{
fill_along(vec)
}
\arguments{
\item{x}{A single integer.}
}
\description{
Fills replace NA values (or specified) with previous non-NA
}
\examples{
x <- c(NA,1,-1,NA, 1, 2,6,4)
max_run(x)
}

