% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fill_run}
\alias{fill_run}
\title{Fill NA with previous non-NA element}
\usage{
fill_run(x, run_for_first = FALSE, only_within = FALSE)
}
\arguments{
\item{x}{A single integer.}
}
\description{
Fill NA with previous non-NA element
}
\examples{
x <- c(NA,1,-1,NA, 1, 2,6,4)
max_run(x)
}

