% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{get_initial_call}
\alias{get_initial_call}
\title{Obtains initial call in recursive chain}
\usage{
get_initial_call(callable)
}
\arguments{
\item{callable}{(\code{argument}) of any function.}
}
\value{
initial (\code{call})
}
\description{
Obtains initial call in recursive chain. Function to be called inside of the
body.
}
