% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mean_run}
\alias{mean_run}
\title{Streak length of vector elements}
\usage{
mean_run(x, n)
}
\arguments{
\item{x}{A single integer.}
}
\description{
Calculates series of consecutive elements
}
\examples{
x <- c(1,1,0,0,1,0,3,3,1,1)
mean_run(x)
}

