% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{minmax_run}
\alias{minmax_run}
\title{Running min/max}
\usage{
minmax_run(x, metric = "min", na_rm = TRUE)
}
\arguments{
\item{x}{Vector of any type}

\item{na_rm}{logical (default \code{na_rm=TRUE}) - if \code{TRUE} sum is calulating excluding \code{NA}.}
}
\value{
list.
}
\description{
\code{min_run} calculates running min on given \code{x} numeric vector, specified \code{k} window size.
}
