% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacer.R
\name{pacer}
\alias{pacer}
\title{Computes target pace based on race time goal}
\usage{
pacer(target.time = 175, race = "marathon")
}
\arguments{
\item{target.time}{n integer value of race time goal (in minutes)}

\item{race}{race type (e.g marathon, 5k, etc)}
}
\value{
target pace
}
\description{
Insert description here
}
