% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_plot.R
\name{progress_plot}
\alias{progress_plot}
\title{Plots the calories burned or heartrate of user's runs over time}
\usage{
progress_plot(name, type = "calories")
}
\arguments{
\item{name}{Name of run log in character vector format. Must be spelled the same way as the logged values in order to access running log}

\item{type}{Type of plot expressed as a character vector. Either 'calories' or 'heartrate' as a function of time are measured. Defualt set is 'calories.}
}
\value{
Returns a plot of type vs. time to give an understanding of progress over time.
}
\description{
Plots the calories burned or heartrate of user's runs over time
}
\examples{
progress_plot("John", type="heartrate")
}

