% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{seq_by}
\alias{seq_by}
\title{Creates sequence for at as time-unit-interval}
\usage{
seq_by(at, idx)
}
\arguments{
\item{at}{object from runner}

\item{idx}{object from runner}
}
\description{
Creates sequence for at as time-unit-interval
}
