% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_log.R
\name{update_log}
\alias{update_log}
\title{Update running log to track running data}
\usage{
update_log(units = "english", date, name, age, gender, height, weight, time,
  distance, heartrate = NA)
}
\arguments{
\item{units}{'english' or 'metric'.}

\item{date}{Date of run. Should be in "mm/dd/yyyy" format.}

\item{name}{User's name. Should be in character format}

\item{age}{User's age. Should be in numeric format.}

\item{gender}{User's gender. Should be in character format, and should either be 'male' or 'female'.}

\item{height}{User's height. Should be in numeric format.}

\item{weight}{User's weight. Should be in numeric format.}

\item{time}{The amount of time the user ran the distance in. Measured in minutes. Should be in character format.}

\item{distance}{The distance the user ran. Measured in miles. Should be in numeric format.}

\item{heartrate}{Optional parameter - user's heartrate during workout. Should be in numeric format.}
}
\value{
Updated dataframe with runner's info.
}
\description{
Update running log to track running data
}
\examples{
update_log("english", "09/30/2016", "Anna", 18, "female", 57, 128, 10, 1)
update_log("metric", "09/30/2016", "Jon", 19, "male", 57, 128, 10, 1, 72)
}

