% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{whichmax_run}
\alias{whichmax_run}
\title{Running which.max}
\usage{
whichmax_run(x, k = 0L, which = "last", na_rm = TRUE, na_pad = FALSE)
}
\arguments{
\item{x}{numeric (or integer) vector for which running whichrun is calculated.}

\item{k}{Running window size. By default window size equals \code{length(x)}. Allow varying window size specified by vector of \code{length(x)}}

\item{which}{specifies whether first or last index is returned.}

\item{na_rm}{logical (default \code{na_rm=TRUE}) - if \code{TRUE} \code{NA} are replaced by last observed minimum prior to element.}

\item{na_pad}{logical (default \code{na_pad=FALSE}) - if \code{TRUE} first k-results will be filled by \code{NA}. If k is not specified na_pad=F by default.}
}
\value{
Numeric vector of length equals length of \code{x} containing running index of maximum in \code{k}-long window.
}
\description{
Running index of the (first or last) maximum.
}
\examples{
x <- c(NA,NA,0,NA, 1, 2,-6,4)
whichmax_run(x)
}

