% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{whichmin_run}
\alias{whichmin_run}
\title{Running length of consecutive occurence}
\usage{
whichmin_run(vec)
}
\arguments{
\item{x}{A single integer.}
}
\description{
Running length of consecutive occurence
}
\examples{
x <- c(NA,NA,99999999999,3.01, 3, 4,-6,-7)
whichmin_run(x)
}

