% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{whicht_run}
\alias{whicht_run}
\title{Running which}
\usage{
whicht_run(x, k = 0L, lag = 0L, which = "last", na_rm = TRUE,
  na_pad = FALSE, idx = integer(0))
}
\arguments{
\item{x}{Vector of any type}

\item{k}{integer vector which specifies window length}

\item{lag}{integer vector which specifies window shift}

\item{na_rm}{logical (default \code{na_rm=TRUE}) - if \code{TRUE} sum is calulating excluding \code{NA}.}

\item{na_pad}{logical (default \code{na_pad=FALSE}) - if \code{TRUE} first k-results will be filled by \code{NA}. If k is not specified na_pad=F by default.}

\item{idx}{an optional integer vector containing index of observations.}
}
\value{
integer vector of indexes of the same length as \code{x}.
}
\description{
\code{min_run} calculates running which - returns index of element where \code{x == TRUE}.
}
\examples{
set.seed(11)
x1 <- sample(c(1, 2, 3), 15, replace = TRUE)
x2 <- sample(c(NA, 1, 2, 3), 15, replace = TRUE)
k  <- sample(1:4, 15, replace = TRUE)
whicht_run(x1)
whicht_run(x2, na_rm = TRUE)
whicht_run(x2, na_rm = TRUE, k = 4)
whicht_run(x2, na_rm = FALSE, k = k)
}
