% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{whichtrue_run}
\alias{whichtrue_run}
\title{Streak length of vector elements}
\usage{
whichtrue_run(vec)
}
\arguments{
\item{x}{A single integer.}
}
\description{
Calculates series of consecutive elements
}
\examples{
x <- c(NA,F,F,T, F, T,F,NA)
whichtrue_run(x)
}

