% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_unresolved_at}
\alias{.check_unresolved_at}
\title{Resolves at argument}
\usage{
.check_unresolved_at(x, at)
}
\arguments{
\item{x}{(\code{data.frame})}

\item{at}{(\code{integer}, \code{Date}, \code{POSIXt}, \code{character} vector)\cr
Vector of any size and any value defining output data points. Values of the
vector defines the indexes which data is computed at. Can be also \code{POSIXt}
sequence increment used in \code{at} argument in \code{\link[base]{seq.POSIXt}}.
See 'Specifying time-intervals' in details section.}
}
\value{
resolved \code{at}
}
\description{
Resolves \code{at} argument passed to the \code{runner} -
checks if the argument has a valid value. If argument is a single character
matching column name in the \code{x} then is replaced with the value \code{x[[at]]}
}
\keyword{internal}
