% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_unresolved_index}
\alias{.check_unresolved_index}
\title{Resolves index argument}
\usage{
.check_unresolved_index(x, idx)
}
\arguments{
\item{x}{(\code{data.frame})}

\item{idx}{(\code{integer}, \code{Date}, \code{POSIXt})\cr
Optional integer vector containing sorted (ascending) index of observation.
By default \code{idx} is index incremented by one. User can provide index with
varying increment and with duplicated values. If specified then \code{k} and
\code{lag} are depending on \code{idx}. Length of \code{idx} have to be equal of length
\code{x}.}
}
\value{
resolved \code{idx}
}
\description{
Resolves \code{at} argument passed to the \code{runner} -
checks if the argument has a valid value. If argument is a single character
matching column name in the \code{x} then is replaced with the value \code{x[[idx]]}
}
\keyword{internal}
