% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fill_run}
\alias{fill_run}
\title{Fill NA with previous non-NA element}
\usage{
fill_run(x, run_for_first = FALSE, only_within = FALSE)
}
\arguments{
\item{x}{Vector of any type where \code{NA} are replaced}

\item{run_for_first}{If first elements are filled with \code{NA}, \code{run_for_first = TRUE } allows to fill all initial \code{NA} with nearest non-NA value. By befault \code{run_for_first = TRUE}}

\item{only_within}{\code{NA} are replaced only if previous and next non-NA values are the same. By befault \code{only_within = TRUE}}
}
\value{
numeric vector of length equals length of \code{x} containing all \code{x} elements with \code{NA} replaced with previous non-NA element.
}
\description{
Fill \code{NA} with last non-NA element.
}
\examples{
x <- c(NA,1,-1,NA, 1, 2,6,4)
max_run(x)
}
