% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{k_by}
\alias{k_by}
\title{Converts k and lag from time-unit-interval to int}
\usage{
k_by(k, idx, param)
}
\arguments{
\item{k}{object from runner}

\item{idx}{object from runner}

\item{param}{name of the parameter to be printed in error message}
}
\description{
Converts k and lag from time-unit-interval to int
}
\examples{
k <-  "1 month"
idx <- seq(as.POSIXct("2019-01-01 03:02:01"), as.POSIXct("2020-01-01 03:02:01"), by = "month")
k_difftime <- runner:::k_by(k, idx, param = "k")
idx - k_difftime
}
