% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lag_run}
\alias{lag_run}
\title{Lag dependent on variable}
\usage{
lag_run(x, lag = 1L, idx = integer(0), nearest = FALSE)
}
\arguments{
\item{x}{(\code{vector}, \code{data.frame}, \code{matrix}, \code{xts}, \code{grouped_df})\cr
input data.}

\item{lag}{(\code{integer} or \code{character})\cr
Window shift. Positive shifts back, negative shifts forward. Single value
or vector of \code{length(x)}. Accepts time-interval strings when \code{idx} is set.}

\item{idx}{(\code{integer}, \code{Date}, \code{POSIXt})\cr
Sorted index of observations. When set, \code{k} and \code{lag} refer to index
distance rather than element count. Must be same length as \code{x}.}

\item{nearest}{\code{logical} single value. Applied when \code{idx} is used,
then \code{nearest = FALSE} returns observation lagged exactly by the
specified number of "periods". When \code{nearest = TRUE}
function returns latest observation within lag window.}
}
\description{
Vector of input lagged along integer vector
}
\examples{
lag_run(1:10, lag = 3)
lag_run(letters[1:10], lag = -2, idx = c(1, 1, 1, 2, 3, 4, 6, 7, 8, 10))
lag_run(letters[1:10], lag = 2, idx = c(1, 1, 1, 2, 3, 4, 6, 7, 8, 10), nearest = TRUE)
}
