% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lag_run}
\alias{lag_run}
\title{Lag dependent on variable}
\usage{
lag_run(x, k = 1L, idx = 1L, nearest = FALSE)
}
\arguments{
\item{x}{Vector of any type}

\item{k}{integer vector which specifies window length}

\item{idx}{an optional integer vector containing index of observations.}

\item{nearest}{logical value. Applied when \code{idx} is used, then \code{nearest = false} returns
observation lagged exactly by the specified number of "periods". When \code{nearest = true}
function returns latest observation within lag window.}
}
\description{
Vector of input lagged along integer vector
}
\examples{
lag_run(1:10, k = 3)
lag_run(letters[1:10], k = 2, idx = c(1, 1, 1, 2, 3, 4, 6, 7, 8, 10))
}
