% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{length_run}
\alias{length_run}
\title{Length of running windows}
\usage{
length_run(k = 1L, idx = 0L)
}
\arguments{
\item{k}{integer vector which specifies window length}

\item{idx}{an optional integer vector containing index of observations.}
}
\description{
Number of elements in k-long window calculated on idx vector. If idx is an `as.integer(date)` vector, then k=number of days in window - then the result is number of observations within k days window.
}
\examples{
length_run(k=3,idx=c(1,2,2,4,5,5,5,5,5,5))
}
