% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{max_run}
\alias{max_run}
\title{Running maximum window}
\usage{
max_run(x, k = 0L, na_rm = TRUE, na_pad = FALSE)
}
\arguments{
\item{x}{input numeric vector where running maximum is calculated.}

\item{k}{Running window size. By default window size equals \code{length(x)}.}

\item{na_rm}{logical (default \code{na_rm=TRUE}) - if \code{TRUE} \code{NA} are replaced by last observed maximum prior to element.}

\item{na_pad}{logical (default \code{na_pad=FALSE}) - if \code{TRUE} first k-results will be filled by \code{NA}. If k is not specified na_pad=F by default.}
}
\value{
numeric vector of length equals length of \code{x} containing running max in \code{k}-long window.
}
\description{
\code{max_run} calculates window-running max on given \code{x} numeric vector, specified \code{k} window size and additional options (\code{na_pad},\code{na_rm}) to handle missing values.
}
\examples{
x <- c(NA,1,-1,NA, -1, -2,-6,4)
max_run(x, k=2, na_rm=TRUE, na_pad=FALSE)
}

