% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{max_run}
\alias{max_run}
\title{Running maximum}
\usage{
max_run(x, k = integer(1), lag = integer(1), na_rm = TRUE,
  na_pad = FALSE, idx = integer(0))
}
\arguments{
\item{x}{Vector of any type}

\item{k}{integer vector which specifies window length}

\item{lag}{integer vector which specifies window shift}

\item{na_rm}{logical (default \code{na_rm=TRUE}) - if \code{TRUE} sum is calulating excluding \code{NA}.}

\item{na_pad}{logical (default \code{na_pad=FALSE}) - if \code{TRUE} first k-results will be filled by \code{NA}. If k is not specified na_pad=F by default.}

\item{idx}{an optional integer vector containing index of observations.}
}
\value{
numeric vector of length equals length of \code{x} containing running max in \code{k}-long window.
}
\description{
\code{min_run} calculates running max on given \code{x} numeric vector, specified \code{k} window size.
}
\examples{
set.seed(11)
x1 <- sample( c(1,2,3), 15, replace=TRUE)
x2 <- sample( c(NA,1,2,3), 15, replace=TRUE)
k  <- sample( 1:4, 15, replace=TRUE)
max_run(x1) # simple cumulative maximum
max_run(x2, na_rm = TRUE) # cumulative maximum with removing NA.
max_run(x2, na_rm = TRUE, k=4) # maximum in 4-element window
max_run(x2, na_rm = FALSE, k=k) # maximum in varying k window size
}
