% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{max_run}
\alias{max_run}
\title{Running maximum window}
\usage{
max_run(x, k = as.integer(c(0)), na_pad = FALSE, na_rm = TRUE)
}
\arguments{
\item{x}{input numeric vector where running maximum is calculated.}

\item{k}{Running window size. By default window size equals the length of x vector argument.}

\item{na_pad}{logical (default=FALSE) - if TRUE first k-results will be filled by NA's. If k is not specified na_pad=F by default.}

\item{na_rm}{logical (default=TRUE) - if TRUE NA's is replaced by last observed maximum.}
}
\description{
Running maximum window
}
\examples{
x <- c(NA,1,-1,NA, -1, -2,-6,4)
max_run(x, k=2, na_rm=TRUE, na_pad=FALSE)
}

