% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mean_run}
\alias{mean_run}
\title{Running mean}
\usage{
mean_run(x, k = 0L, na_rm = TRUE, na_pad = FALSE)
}
\arguments{
\item{x}{vector of any type on which running mean is calculated}

\item{k}{running window size. By default window size equals \code{length(x)}. Allow varying window size specified by vector of \code{length(x)}}

\item{na_rm}{logical (default \code{na_rm=TRUE}) - if \code{TRUE} mean is calulating excluding \code{NA}.}

\item{na_pad}{logical (default \code{na_pad=FALSE}) - if \code{TRUE} first k-results will be filled by \code{NA}. If k is not specified na_pad=F by default.}
}
\value{
numeric vector of length equals length of \code{x} containing running mean in \code{k}-long window.
}
\description{
Running mean
}
\examples{
x <- c(NA,3,4,3,5,7,NA,2,4,9)
mean_run(x , k = 3, na_rm = TRUE, na_pad = FALSE)
}
