% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{min_run}
\alias{min_run}
\title{Running maximum window}
\usage{
min_run(x, k = 0L, na_pad = FALSE, na_rm = TRUE)
}
\arguments{
\item{x}{input numeric vector where running minimum is calculated.}

\item{k}{Running window size. By default window size equals the length of \code{x}.}

\item{na_pad}{logical (default \code{na_pad=FALSE}) - if \code{TRUE} first k-results will be filled by \code{NA}. If k is not specified na_pad=F by default.}

\item{na_rm}{logical (default \code{na_rm=TRUE}) - if \code{TRUE} \code{NA} are replaced by last observed minimum prior to element.}
}
\value{
numeric vector of length equals length of \code{x} containing running min in \code{k}-long window.
}
\description{
\code{min_run} calculates window-running max on given \code{x} numeric vector, specified \code{k} window size and additional options (\code{na_pad},\code{na_rm}) to handle missing values.
}
\examples{
x <- c( NA, 1, -1, NA, -1, -2, -6, 4)

min_run( x, k=2, na_rm=TRUE, na_pad=FALSE )
}

