% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{minmax_run}
\alias{minmax_run}
\title{Running min/max}
\usage{
minmax_run(x, metric = "min", na_rm = TRUE)
}
\arguments{
\item{x}{(\code{vector}, \code{data.frame}, \code{matrix}, \code{xts})\cr
Input in runner custom function \code{f}.}

\item{metric}{\code{character} what to return, minimum or maximum}

\item{na_rm}{\code{logical} single value (default \code{na_rm = TRUE}) -
if \code{TRUE} sum is calculating excluding \code{NA}.}
}
\value{
list.
}
\description{
\code{min_run} calculates running minimum-maximum on given \code{x} numeric
 vector, specified \code{k} window size.
}
