% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runner}
\alias{runner}
\title{Streak length of vector elements}
\usage{
runner(x, k = 0L, lag = 0L)
}
\arguments{
\item{x}{A single integer.}
}
\description{
Calculates series of consecutive elements
}
\examples{
x <- c(1,3,4,3,5,7,8,2,4,9)
mean_run(x)
}

