% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{streak_run}
\alias{streak_run}
\title{Streak length of vector elements}
\usage{
streak_run(x, k = 0L, na_rm = FALSE, na_pad = FALSE)
}
\arguments{
\item{x}{Vector of any type on which consecutive streak is calculated}

\item{k}{Running window size. By default window size equals \code{length(x)}.}

\item{na_rm}{logical (default \code{na_rm=TRUE}) - if \code{TRUE} \code{NA} are replaced by last observed streak prior to element.}

\item{na_pad}{logical (default \code{na_pad=FALSE}) - if \code{TRUE} first k-results will be filled by \code{NA}. If k is not specified na_pad=F by default.}
}
\value{
numeric vector of length equals length of \code{x} containing running streak length in \code{k}-long window.
}
\description{
Calculates series of consecutive elements
}
\examples{
x <- c(1,1,0,0,1,0,3,3,1,1)
streak_run(x, k=2, na_pad=FALSE)
}

