% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{streak_run}
\alias{streak_run}
\title{Streak length of vector elements}
\usage{
streak_run(vec, k = NA_integer_)
}
\arguments{
\item{x}{A single integer.}
}
\description{
Calculates series of consecutive elements
}
\examples{
x <- c(1,1,0,0,1,0,3,3,1,1)
streak_run(x)
}

