% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sum_run}
\alias{sum_run}
\title{Running window sum}
\usage{
sum_run(x, k = 0L, na_rm = TRUE, na_pad = FALSE)
}
\arguments{
\item{x}{Vector of any type on which running sum is calculated}

\item{k}{Running window size. By default window size equals \code{length(x)}.}

\item{na_rm}{logical (default \code{na_rm=TRUE}) - if \code{TRUE} sum is calulating excluding \code{NA}.}

\item{na_pad}{logical (default \code{na_pad=FALSE}) - if \code{TRUE} first k-results will be filled by \code{NA}. If k is not specified na_pad=F by default.}
}
\value{
numeric vector of length equals length of \code{x} containing running window sum in \code{k}-long window.
}
\description{
Running window sum
}
\examples{
x <- c(NA,3,4,3,5,7,NA,2,4,9)
sum_run(x , k = 3, na_rm = T, na_pad = F)
}

