% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{whichmax_run}
\alias{whichmax_run}
\title{Streak length of vector elements}
\usage{
whichmax_run(vec)
}
\arguments{
\item{x}{A single integer.}
}
\description{
Calculates series of consecutive elements
}
\examples{
x <- c(NA,NA,0,NA, 1, 2,-6,4)
whichmax_run(x)
}

