% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{whichtrue_run}
\alias{whichtrue_run}
\title{Running index of last TRUE element}
\usage{
whichtrue_run(vec)
}
\arguments{
\item{x}{An logical vector}
}
\description{
Running index of last TRUE element
}
\examples{
x <- c(NA,FALSE,FALSE,TRUE, FALSE, TRUE,FALSE,NA)
whichtrue_run(x)
}

