% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{window_run}
\alias{window_run}
\title{List of running windows}
\usage{
window_run(x, k = 0L)
}
\arguments{
\item{x}{Vector of any type}

\item{k}{integer vector which specifies window length}
}
\description{
Creates list of windows
}
\examples{
window_run(1:10, k=3)
window_run(letters[1:10],k=c(1,2,2,4,5,5,5,5,5,5))
}
